feat: добавлен функционал бронирования и отзывов

+++ src/main/java/ru/practicum/shareit/user/dto/UserDto.java
@@ -9,7 +9,7 @@ import lombok.experimental.FieldDefaults;
 @FieldDefaults(level = AccessLevel.PRIVATE)
 @Builder
 public class UserDto {
-    int id;
+    Long id;
     String name;
     String email;
 }
